refactor: 거리 생성 메서드 수정
