+++ src/test/java/subway/fixture/DistanceFixture.java
@@ -1,10 +1,10 @@
 package subway.fixture;
 
-import subway.domain.Distance;
+import subway.domain.vo.Distance;
 
 public abstract class DistanceFixture {
 
     public static Distance 거리(final int value) {
-        return new Distance(value);
+        return Distance.from(value);
     }
 }
